@@ -1,4 +1,4 @@
-declare module "*.jpg";
-declare module "*.png";
-declare module "*.jpeg";
-declare module "*.gif";
+declare module '*.jpg'
+declare module '*.png'
+declare module '*.jpeg'
+declare module '*.gif'
