@@ -1,9 +1,8 @@
-import { AbstractEmployee } from './AbstractEmployee'
-import { DeveloperSkills } from '../model/employeeModel'
 import { LimitedNumber, normalizedRandom } from '@potato-golem/core'
+import { DeveloperSkills } from '../model/employeeModel'
+import { AbstractEmployee } from './AbstractEmployee'
 
 export class DeveloperEmployee extends AbstractEmployee<DeveloperSkills> {
-
   constructor() {
     super()
 
