@@ -1,14 +1,14 @@
 import { Game } from 'phaser'
-import { MainMenuScene } from './scenes/MainMenuScene'
-import RexUIPlugin from 'phaser3-rex-plugins/templates/ui/ui-plugin'
 import RexImageBoxPlugin from 'phaser3-rex-plugins/plugins/imagebox-plugin.js'
+import RexUIPlugin from 'phaser3-rex-plugins/templates/ui/ui-plugin'
+import { MainMenuScene } from './scenes/MainMenuScene'
 import { BoardScene } from './scenes/board/BoardScene'
 
 const GameResolutions = {
   default: {
     width: 1280,
     height: 720,
-  }
+  },
 }
 
 const resolution = GameResolutions.default
@@ -19,23 +19,22 @@ const config: Phaser.Types.Core.GameConfig = {
   width: resolution.width,
   height: resolution.height,
   plugins: {
-    global: [{
-      key: 'rexImageBoxPlugin',
-      plugin: RexImageBoxPlugin,
-      start: true
-    },
-      ],
-    scene: [{
-      key: 'rexUI',
-      plugin: RexUIPlugin,
-      mapping: 'rexUI',
-    },
+    global: [
+      {
+        key: 'rexImageBoxPlugin',
+        plugin: RexImageBoxPlugin,
+        start: true,
+      },
+    ],
+    scene: [
+      {
+        key: 'rexUI',
+        plugin: RexUIPlugin,
+        mapping: 'rexUI',
+      },
     ],
   },
-  scene: [
-    new MainMenuScene(),
-    new BoardScene(),
-  ],
+  scene: [new MainMenuScene(), new BoardScene()],
 }
 
 // For some reason this must happen in root TS file
