@@ -1,48 +1,48 @@
 export enum Gender {
-  male = "male",
-  female = "female",
+  male = 'male',
+  female = 'female',
 }
 
 export const DirectorPortraits = {
   [Gender.male]: [
-    "male1",
-    "male2",
-    "male3",
-    "male4",
-    "male5",
-    "male6",
-    "male7",
-    "male8",
-    "male9",
-    "male10",
-    "male11",
-    "male12",
-    "male13",
-    "male14",
-    "male15",
-    "male16",
-    "male17",
-    "male18",
-    "male19",
-    "male20",
+    'male1',
+    'male2',
+    'male3',
+    'male4',
+    'male5',
+    'male6',
+    'male7',
+    'male8',
+    'male9',
+    'male10',
+    'male11',
+    'male12',
+    'male13',
+    'male14',
+    'male15',
+    'male16',
+    'male17',
+    'male18',
+    'male19',
+    'male20',
   ],
   [Gender.female]: [
-    "female1",
-    "female2",
-    "female3",
-    "female4",
-    "female5",
-    "female6",
-    "female7",
-    "female8",
-    "female9",
-    "female10",
-    "female11",
-    "female12",
-    "female13",
-    "female14",
-    "female15",
-    "female16",
-    "female17",
+    'female1',
+    'female2',
+    'female3',
+    'female4',
+    'female5',
+    'female6',
+    'female7',
+    'female8',
+    'female9',
+    'female10',
+    'female11',
+    'female12',
+    'female13',
+    'female14',
+    'female15',
+    'female16',
+    'female17',
   ],
-};
+}
