@@ -1,9 +1,9 @@
 export enum Scenes {
-  MAIN_MENU_SCENE = "mainMenu",
-  BOARD_SCENE = "board",
-  OVERVIEW_SCENE = "overview",
+  MAIN_MENU_SCENE = 'mainMenu',
+  BOARD_SCENE = 'board',
+  OVERVIEW_SCENE = 'overview',
 
-  TURN_RESULTS = "turnResults",
-  HIRE_DIRECTOR = "hireDirector",
-  EXECUTIVE_COUNCIL = 'executiveCouncil'
+  TURN_RESULTS = 'turnResults',
+  HIRE_DIRECTOR = 'hireDirector',
+  EXECUTIVE_COUNCIL = 'executiveCouncil',
 }
