@@ -3,5 +3,5 @@ import { DEFAULT_ENTITY_TYPE } from '@potato-golem/ui'
 export const EntityTypeRegistry = {
   DEFAULT: DEFAULT_ENTITY_TYPE,
   TICKET: 'ticket',
-  ENGINEER: 'engineer'
+  ENGINEER: 'engineer',
 }
