@@ -1,13 +1,14 @@
-import { Scenes } from '../registries/SceneRegistry'
 import {
   BackgroundBuilder,
   ButtonListBuilder,
   ChangeSceneActivation,
-  MultiplexActivation, NameValueLabelBuilder,
+  MultiplexActivation,
+  NameValueLabelBuilder,
   UIGroupSlot,
 } from '@potato-golem/ui'
 import { PotatoScene } from '@potato-golem/ui/dist/src/ui/common/PotatoScene'
 import { COLOR_DARK, COLOR_LIGHT } from '@potato-golem/ui/dist/src/ui/constants/Colours'
+import { Scenes } from '../registries/SceneRegistry'
 
 export class OverviewScene extends PotatoScene {
   constructor() {
