@@ -1,7 +1,7 @@
-import { UIElementTemplate } from "@potato-golem/ui";
+import { UIElementTemplate } from '@potato-golem/ui'
 
 export const buttonTemplate: UIElementTemplate = {
   textureKey: 'glass-panel',
   displaySizeX: 150,
-  displaySizeY: 50
+  displaySizeY: 50,
 }
