@@ -1,8 +1,8 @@
-import { AbstractUIBuilder } from './AbstractUIBuilder'
-import { COLOR_DARK, COLOR_LIGHT } from '../constants/Colours'
-import { PotatoScene } from '../common/PotatoScene'
-import { ImageBoxBuilder } from './ImageBoxBuilder'
 import { validateNumber } from 'validation-utils'
+import { PotatoScene } from '../common/PotatoScene'
+import { COLOR_DARK, COLOR_LIGHT } from '../constants/Colours'
+import { AbstractUIBuilder } from './AbstractUIBuilder'
+import { ImageBoxBuilder } from './ImageBoxBuilder'
 
 export type RowDefinition = {
   icon?: string
@@ -15,9 +15,7 @@ export type RowDefinition = {
 }
 
 export class SizerBuilder extends AbstractUIBuilder {
-
   private rowDefinitions: RowDefinition[]
-
 
   constructor(scene: PotatoScene) {
     super(scene)
@@ -25,29 +23,30 @@ export class SizerBuilder extends AbstractUIBuilder {
   }
 
   addRow(row: RowDefinition) {
-    this.rowDefinitions.push({...row})
+    this.rowDefinitions.push({ ...row })
     return this
   }
 
   build() {
-    const rowCount = 3; // Change it to 3, 5, 10
+    const rowCount = 3 // Change it to 3, 5, 10
 
     const sizer = this.scene.rexUI.add.sizer({
-      x: this.getX(), y: this.getY(),
-      width: this.getWidth(), height: this.getHeight(),
-      orientation: 1
+      x: this.getX(),
+      y: this.getY(),
+      width: this.getWidth(),
+      height: this.getHeight(),
+      orientation: 1,
     })
 
-    for (let row of this.rowDefinitions) {
-      sizer.add(
-        this.createLabel(this.scene, row),
-        { proportion: row.proportion ?? 1, expand: true,
+    for (const row of this.rowDefinitions) {
+      sizer.add(this.createLabel(this.scene, row), {
+        proportion: row.proportion ?? 1,
+        expand: true,
         //  align: row.icon ? 'center-center' : 'center-center'
-        }
-      )
+      })
     }
 
-    sizer.layout();
+    sizer.layout()
   }
 
   createLabel(scene: PotatoScene, row: RowDefinition) {
@@ -66,15 +65,20 @@ export class SizerBuilder extends AbstractUIBuilder {
     }
 
     return scene.rexUI.add.label({
-      background: scene.rexUI.add.roundRectangle(0, 0, 0, 0, 0, COLOR_DARK, row.alpha ?? 1).setDepth(0), //{ strokeColor: COLOR_LIGHT }),
+      background: scene.rexUI.add
+        .roundRectangle(0, 0, 0, 0, 0, COLOR_DARK, row.alpha ?? 1)
+        .setDepth(0), //{ strokeColor: COLOR_LIGHT }),
       text: row.text ? scene.add.text(0, 0, row.text).setWordWrapWidth(this.getWidth()) : undefined,
       icon: icon ? icon : undefined,
       squareFitIcon: true,
       space: {
-        left: row.iconOffsetX ? row.iconOffsetX : 10, right: 10, top: 10, bottom: 10,
-        icon: 10, text: 10,
-      }
+        left: row.iconOffsetX ? row.iconOffsetX : 10,
+        right: 10,
+        top: 10,
+        bottom: 10,
+        icon: 10,
+        text: 10,
+      },
     })
   }
-
 }
