@@ -1,5 +1,5 @@
 import { Scene } from 'phaser'
-import UIPlugin from 'phaser3-rex-plugins/templates/ui/ui-plugin.js';
+import UIPlugin from 'phaser3-rex-plugins/templates/ui/ui-plugin.js'
 
 export class PotatoScene extends Scene {
   rexUI!: UIPlugin
