@@ -1,10 +1,10 @@
-import { AbstractUIElementLite } from './UIGroup'
 import { ENTITY_MODEL, ENTITY_TYPE_DATA_KEY } from '../common/EntityDataKeys'
+import { AbstractUIElementLite } from './UIGroup'
 
 export const DEFAULT_ENTITY_TYPE = 'DEFAULT'
 
 export function storeStartPosition(item: AbstractUIElementLite) {
-  item.setData({ startX: item.x, startY: item.y });
+  item.setData({ startX: item.x, startY: item.y })
 }
 
 export function restoreStartPosition(item: AbstractUIElementLite) {
